PLSR-313: ClientImpl::handleClose shouldn't use statics
ClientImpl::handleClose was using static variables to record the first
error on closing a client. This is just wrong. A static stack variable
in c++ acts like a global. So if errorClosing was ever set to true in
a process, all clients closed in that process after that point would
be errored with first error.

This fixes a failure in BasicEndToEndTest.testDelayedMessages, which
happens sporadically on C++, and always when running the tests in
serial, probably due to a double close in some other test.