PLSR-254: Reenable ServerCnxTest#testConnectCommandWithAuthenticationPositive
The test disabled even though the failure happened in a dev branch
rather than the master branch. The cause of the failure was fixed by the
time the change made it to master so there's nothing to fix.

The issue was caused by ServerCnx checking whether a authRole is null
when doing authentication. ServerCnxTest sets it to "". For the
failure in question, on toAuthentication, ServerCnx sees authRole as
non-null and closes the channel, failing the test. The final change
checks for null or empty so it passes correctly.