PLSR-401: MockZooKeeper shouldn't 'execute' while holding lock
There was one particular bad code path where MockZooKeeper#delete
acquires the state mutex, submits to the executor and then releases
the lock. Normally this would be fine. However, MockZooKeeper is often
constructed with a direct executor. This means that the lock is still
held by the delete when it's acquired in the executed code block. So
even if we release before calling callbacks, the lock will still be
held until the end of the #delete() call.